feat: removed plugin system
