+++ src/foremast/app/__init__.py
@@ -14,4 +14,4 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 """Package for creating a Spinnaker application."""
-from .create_app import SpinnakerApp
+from .spinnaker_app import SpinnakerApp
