+++ src/foremast/templates/infrastructure/app_data.json.j2
@@ -3,7 +3,7 @@
     {
         "type":"createApplication",
         "application":{
-            "cloudProviders":"aws",
+            "cloudProviders":"{{ appinfo.provider }}",
             "name": "{{ appinfo.app }}",
             "email": "{{ appinfo.email }}",
             "repoType": "github",
